requirements.txt
alembic==1.13.2
authlib==1.3.1 
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
deep-translator==1.11.4
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-socketio==5.3.6
geographiclib==2.0
geopy==2.4.1
gunicorn==22.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
hyperframe==5.2.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
opencage==2.4.0
packaging==24.1
pipdeptree==2.23.0
psycopg2-binary==2.9.9
pycountry==23.12.11
pydexcom==0.4.0
python-dotenv==0.19.2
py-googletrans==1.2
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3

